+++ backend/src/main/java/com/techcoders/petsitter/security/services/UserDetailsImpl.java
@@ -17,7 +17,7 @@ public class UserDetailsImpl implements UserDetails{
 
 	private static final long serialVersionUID = 1L;
 	
-	private Long id;
+	private Integer id;
 	
 	private String username;
 	
@@ -34,7 +34,7 @@ public class UserDetailsImpl implements UserDetails{
 	
 	private Collection<? extends GrantedAuthority> authorities;
 
-	public UserDetailsImpl(Long id, String username, String cpf, String phone, Double rating, String email, String password,
+	public UserDetailsImpl(Integer id, String username, String cpf, String phone, Double rating, String email, String password,
 			Collection<? extends GrantedAuthority> authorities) {
 		this.id = id;
 		this.username = username;
@@ -69,7 +69,7 @@ public class UserDetailsImpl implements UserDetails{
 		return authorities;
 	}
 		
-	public Long getId() {
+	public Integer getId() {
 		return id;
 	}
